% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_build.R
\name{build_scraper}
\alias{build_scraper}
\title{generates the scraping function defined by start_scraper and other constructors in your environment}
\usage{
build_scraper(prev = NULL)
}
\arguments{
\item{prev}{a placeholder for the output of functions being piped into show(). Defaults to NULL and should not be altered.}
}
\value{
a function
}
\description{
generates the scraping function defined by start_scraper and other constructors in your environment
}
\examples{
\dontrun{

start_scraper(args = c("x"), name = "fun") \%>>\%
go("x") \%>>\%
build_scraper()


}
}
