% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_rselenium.R
\name{close_rselenium}
\alias{close_rselenium}
\title{utility function that closes all parallel instances of rselenium}
\usage{
close_rselenium()
}
\description{
utility function that closes all parallel instances of rselenium
}
