% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_elements.R
\name{gen_varname}
\alias{gen_varname}
\title{utility function to check for repeated and generate unique variable names}
\usage{
gen_varname(input)
}
\arguments{
\item{input}{character string}
}
\value{
generated variable name as a character string
}
\description{
utility function to check for repeated and generate unique variable names
}
\keyword{internal}
