% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_navigation.R
\name{go}
\alias{go}
\title{wrapper around \code{RSelenium remDr$mavigate} to generate safe navigation code}
\usage{
go(url)
}
\value{
a character string defining 'RSelenium' navigation instructions that can be pasted into a scraping function
}
\description{
|@param url a character string specifying the name of the variable holding the url string or the url string the function should navigate to.
}
\examples{

go("https://www.wikipedia.org/")
}
