% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_navigation.R
\name{goback}
\alias{goback}
\title{wrapper around \code{RSelenium remDr$goBack} to generate safe backwards navigation code}
\usage{
goback()
}
\value{
a character string defining 'RSelenium' backwards navigation instructions that can be pasted into a scraping function
}
\description{
wrapper around \code{RSelenium remDr$goBack} to generate safe backwards navigation code
}
\examples{

goback()
}
