% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_navigation.R
\name{goforward}
\alias{goforward}
\title{wrapper around remDr$goForward method to generate safe forwards navigation code}
\usage{
goforward(prev = NULL)
}
\arguments{
\item{prev}{a placeholder for the output of functions being piped into goforward(). Defaults to NULL and should not be altered.}
}
\value{
a character string defining 'RSelenium' forward navigation instructions that can be pasted into a scraping function.
}
\description{
wrapper around remDr$goForward method to generate safe forwards navigation code
}
\examples{
\dontrun{

goforward() \%>>\%
show()

}
}
