% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_pipe.R
\name{\%>>\%}
\alias{\%>>\%}
\title{pipe-like operator that pastes together the output of two functions}
\usage{
step_1 \%>>\% step_2
}
\arguments{
\item{step_1}{a function call returning a charater string}

\item{step_2}{a fuction call returning a character string which should be pasted to \code{step_1}}
}
\value{
a character string of the pasted output of \code{step_1} and \code{step_2}
}
\description{
pipe-like operator that pastes together the output of two functions
}
\examples{

go("https://www.wikipedia.org/") \%>>\%
goback()
}
