% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parscrape.R
\name{parscrape}
\alias{parscrape}
\title{parallelize execution of RSelenium}
\usage{
parscrape(
  scrape_fun,
  scrape_input,
  cores,
  packages,
  browser,
  ports = NULL,
  chunk_size = 10
)
}
\arguments{
\item{scrape_fun}{a function with input x sending instructions to remDr (remote driver)/ scraping function to be parallelized}

\item{scrape_input}{a data frame, list, or vector where each element is an input to be passed to scrape_function}

\item{cores}{number of cores to run RSelenium instances on}

\item{packages}{a character vector with package names of packages used in scrape_function}

\item{browser}{a character vector specifying the browser to be used}

\item{ports}{vector of ports for RSelenium instances (if left at default NULL parscrape will randomly generate ports)}

\item{chunk_size}{number of scrape_input elements to be processed per round of scrape_function (parscrape splits scrape_input into chunks and runs scrape_function in multiple rounds to avoid loosing data due to errors)}
}
\value{
output of scrape_function
}
\description{
parallelize execution of RSelenium
}
