% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_build.R
\name{show}
\alias{show}
\title{renders the output of the piped functions to the console via cat()}
\usage{
show(prev = NULL)
}
\arguments{
\item{prev}{a placeholder for the output of functions being piped into show(). Defaults to NULL and should not be altered.}
}
\value{
None (invisible NULL)
}
\description{
renders the output of the piped functions to the console via cat()
}
\examples{
\dontrun{

go("https://www.wikipedia.org/") \%>>\%
goback() \%>>\%
show()

}
}
