% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_build.R
\name{start_scraper}
\alias{start_scraper}
\title{sets function name and arguments of scraping function}
\usage{
start_scraper(args, name = NULL)
}
\arguments{
\item{args}{a character vector of function arguments}

\item{name}{character string specifying the object name of the scraping function. If NULL defaults to 'scraper'}
}
\value{
a character string starting a function definition
}
\description{
sets function name and arguments of scraping function
}
\examples{
\dontrun{

start_scraper(args = c("x","y"), name = "fun")

}
}
