% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_rselenium.R
\name{close_rselenium}
\alias{close_rselenium}
\title{utility function that closes all parallel instances of RSelenium}
\usage{
close_rselenium(clust)
}
\arguments{
\item{clust}{\code{parallel} cluster}
}
\value{
No return value, called to close RSelenium instances in parscrape.
}
\description{
utility function that closes all parallel instances of RSelenium
}
\keyword{internal}
