% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_navigation.R
\name{goback}
\alias{goback}
\title{wrapper around remDr$goBack method to generate safe backwards navigation code}
\usage{
goback(prev = NULL)
}
\arguments{
\item{prev}{a placeholder for the output of functions being piped into goback(). Defaults to NULL and should not be altered.}
}
\value{
a character string defining 'RSelenium' backwards navigation instructions that can be pasted into a scraping function
}
\description{
wrapper around remDr$goBack method to generate safe backwards navigation code
}
\examples{
\dontrun{

goback() \%>>\%
show()

}
}
