% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_pipe.R
\name{\%>>\%}
\alias{\%>>\%}
\title{pipe-like operator that pastes together the output of two functions}
\usage{
lhs \%>>\% rhs
}
\arguments{
\item{lhs}{a function call returning a charater string}

\item{rhs}{a fuction call returning a character string which should be pasted to step_1}
}
\value{
a character string of the pasted output of step_1 and step_2
}
\description{
pipe-like operator that pastes together the output of two functions
}
\examples{
\dontrun{

go("https://www.wikipedia.org/") \%>>\%
goback()

}
}
