% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors_pipe.R
\name{\%>>\%}
\alias{\%>>\%}
\title{pipe-like operator that passes the output of lhs to the prev argument of rhs to paste together a scraper function in sequence.}
\usage{
lhs \%>>\% rhs
}
\arguments{
\item{lhs}{a parsel constructor function call}

\item{rhs}{a parsel constructor function call that should accept lhs as its prev argument}
}
\value{
the output of rhs evaluated with lhs as the prev argument
}
\description{
pipe-like operator that passes the output of lhs to the prev argument of rhs to paste together a scraper function in sequence.
}
\examples{
\dontrun{

#paste together the go and goback output in sequence
go("https://www.wikipedia.org/") \%>>\%
goback()

}
}
