% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsel-package.R
\docType{package}
\name{parsel-package}
\alias{parsel}
\alias{parsel-package}
\title{parsel: Prallelized Web Scraping With Rselenium}
\description{
Speed up your RSelenium webscraping by scraping on multiple browsers 
    simultaniously.
}
\keyword{internal}
