% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsel-package.R
\docType{package}
\name{parsel-package}
\alias{parsel}
\alias{parsel-package}
\title{parsel: Parallel Execution of RSelenium}
\description{
Speed up RSelenium Web scraping by scraping on multiple browsers 
    simultaneously.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/till-tietz/parsel}
  \item Report bugs at \url{https://github.com/till-tietz/parsel/issues}
}

}
\author{
\strong{Maintainer}: Till Tietz \email{ttietz2014@gmail.com}

}
\keyword{internal}
