% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsel-package.R
\docType{package}
\name{parsel-package}
\alias{parsel}
\alias{parsel-package}
\title{parsel: Parallelized Dynamic Web-Scraping Using 'RSelenium'}
\description{
A system to increase the efficiency of dynamic web-scraping with 'RSelenium'
    by leveraging parallel processing. You provide a function wrapper for your 'RSelenium' 
    scraping routine with a set of inputs, and 'parsel' runs it in several browser instances. 
    Chunked input processing as well as error catching and logging ensures seamless 
    execution and minimal data loss, even when unforeseen 'RSelenium' errors occur.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/till-tietz/parsel}
  \item Report bugs at \url{https://github.com/till-tietz/parsel/issues}
}

}
\author{
\strong{Maintainer}: Till Tietz \email{ttietz2014@gmail.com}

}
\keyword{internal}
